requirements.txt
keras==2.13.1
matplotlib==3.7.5
numpy==1.24.4
opt_einsum==3.3.0
scipy==1.13.1
skimage==0.0
tensorflow_cpu==2.13.1
